+++ Model/LookUpModel.cs
@@ -1,14 +1,11 @@
 ﻿using Ascon.Pilot.SDK;
-using PilotLookUp.Core;
 using PilotLookUp.Extensions;
-using PilotLookUp.Model.Utils;
 using PilotLookUp.Objects;
 using PilotLookUp.Objects.TypeHelpers;
 using PilotLookUp.Utils;
 using PilotLookUp.ViewModel;
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -49,7 +46,8 @@ namespace PilotLookUp.Model
             return res;
         }
 
-        public async Task<ObjectSet> SearchByString(string request) => await SearchUtils.ByString(_objectsRepository, request);
+        public async Task<ObjectSet> SearchByString(string request)
+            => await SearchUtils.GetObjByString(_objectsRepository, request);
 
 
         public LookUpVM GetDBLookUpVM()
@@ -68,101 +66,17 @@ namespace PilotLookUp.Model
         }
 
         public void GoTo(IDataObject dataObject)
-        {
-            _tabServiceProvider.ShowElement(dataObject.Id);
-        }
+           => _tabServiceProvider.ShowElement(dataObject.Id);
 
-        public async Task<DataObjectHelper> FindLastParentHelper(IDataObject dataObject)
-        {
-            if (dataObject != null)
-            {
-                var parent = await dataObject.FindLastParrent(_objectsRepository);
-                var objSet = await new Tracer(_objectsRepository, null, null).Trace(parent);
-                return objSet.FirstOrDefault() is DataObjectHelper result ? result : null;
-            }
-            return null;
-        }
 
-        public async Task<ListItemVM> FillChild(ListItemVM lastParrent) => await TreeViewUtils.FillChild(_objectsRepository,lastParrent);
-        //{
-        //    await BuildChildNodes(lastParrent);
-        //    return lastParrent;
-        //}
+        public async Task<ListItemVM> FillChild(ListItemVM lastParrent)
+            => await TreeViewUtils.FillChild(_objectsRepository, lastParrent);
 
-        //public async Task BuildChildNodes(ListItemVM lastParrent)
-        //{
-        //    var sad = lastParrent.PilotObjectHelper.LookUpObject as IDataObject;
-        //    List<Guid> children = sad.Children.ToList();  // Метод получения детей по ID
-        //    ObjectSet newPilotObj = await new Tracer(_objectsRepository, null, null).Trace(children);
-        //    foreach (var dataObjectHelper in newPilotObj)
-        //    {
-        //        var childNode = new ListItemVM(dataObjectHelper);
-        //        if (lastParrent.Children != null)
-        //        {
-        //            lastParrent.Children.Add(childNode);
-        //        }
-        //        else
-        //        {
-        //            lastParrent.Children = new ObservableCollection<ListItemVM>()
-        //            {
-        //                childNode
-        //            };
-        //        }
-        //        await BuildChildNodes(childNode); // Рекурсия для вложенных детей
-        //    }
-        //}
-        //internal async Task<List<IDataObject>> GetChildrensWithTimeout(IObjectsRepository objectsRepository, IDataObject currentObject, int timeoutMilliseconds = 300)
-        //{
-        //    var loader = new ObjectLoader(objectsRepository);
-        //    var childrensId = currentObject.Children;
-        //    List<IDataObject> childrens = new List<IDataObject>();
-        //    foreach (var child in childrensId)
-        //    {
-        //        var childObj = await loader.LoadWithTimeout(child);
-        //        childrens.Add(childObj);
-        //    }
-        //    return childrens;
-        //}
 
-        /// <summary>
-        /// По карточке объекта ищет всязи связи с типо задания. Если задание в процессе то добавляет уникальный процесс, а не задания
-        /// </summary>
-        /// <param name="objectHelper"></param>
-        /// <returns></returns>
-        /// <exception cref="NotImplementedException"></exception>
-        internal async Task<ObjectSet> FindAllLastParrent(PilotObjectHelper objectHelper, bool findRevoked = false)
-        {
-            ObjectSet pilotObjectHelpers = new ObjectSet(null);
-            ObjectSet childrenSet;
+        public async Task<DataObjectHelper> SearchLastParent(IDataObject dataObject)
+           => await SearchUtils.GetLastParent(_objectsRepository, dataObject);
 
-            if (objectHelper.LookUpObject is IDataObject dataObject)
-            {
-                childrenSet = await new Tracer(_objectsRepository, null, null).Trace(dataObject.Children);
-            }
-            else return null;
-
-            foreach (var child in childrenSet)
-            {
-                if (child is DataObjectHelper dataHelp)
-                {
-                    if (dataHelp.IsTask)
-                    {
-                        if (!findRevoked) // Пропуск отозванных заданий
-                        {
-                            if (dataHelp.IsRevokedTask)
-                            {
-                                continue;
-                            }
-                        }
-                        DataObjectHelper lastParrent = await FindLastParentHelper(dataObject);
-                        if (!pilotObjectHelpers.Select(it => it.StringId).Contains(lastParrent.StringId))
-                        {
-                            pilotObjectHelpers.Add(lastParrent);
-                        }
-                    }
-                }
-            }
-            return pilotObjectHelpers;
-        }
+        public async Task<ObjectSet> SearchBaseParentsOfChildren(PilotObjectHelper objectHelper, bool findRevoked = false)
+            => await SearchUtils.GetBaseParentsOfChildren(_objectsRepository, objectHelper, findRevoked);
     }
 }
